Motor Mixer Finished
Removed Servo Library and created my own methods to control the weapon's ESC using millis() to create the desired waveform for the corresponding speed. Using the Servo library utilizes the pwm clock used for pins 9 and 10, of which pin 9 is used to control on of the directions of a drive motor. Fixed some issues with residual voltage (~70mV) being on the output of the motors by turning off the enable pin when the inputs are near 0 (within the deadband specified).

Refactored some of the main loop to maximize the loop speed so the "crude" method of using millis() to create the desired digital waveform for the weapon's ESC signal (so there is no delay between when the signal should go low, and when the loop reaches to code to set it low).